Lua-coerced World instance (it's now possible to script most of in-game things using lua)
+ Test init new location lua script
+ Overlay refresh fix
+ MapObject harvestHp typo fix

Content:
+New Quest (tutorial branch) -- find a dungeon entrance & kill 10 rats
+ Coin weight is now 0
+ Wood weight reduced to 0.25 & Max inventory weight multiplier increased to 7.5 (because sometimes it was impossible to complete the first tutorial quest... meh)

+++ MadSand-FallenRiseFromDust-core/src/hitonoriol/madsand/entities/inventory/Inventory.java
@@ -36,7 +36,7 @@ public class Inventory {
 		itemUI = new HashMap<Item, InventoryUICell>();
 	}
 
-	public void setMaxWeight(float val) {
+	public void setMaxWeight(double val) {
 		maxWeight = val;
 		refreshUITitle();
 	}
